fix: keep metamask log in after lit logout

+++ src/utils/Helpers.ts
@@ -121,27 +121,12 @@ const isLitLogin = (val: string) => {
 
 const checkPreviousLoginMode = (account: string) => {
     const queryParams = new URLSearchParams(location.search);
-    const allKeys = Object.keys(localStorage);
-
-    const keysToKeep = allKeys.filter(key => key.startsWith('clientID') || key.startsWith('streamID'));
-
-    const keysAndValues: Record<string, string> = {};
-
-    keysToKeep.forEach(key => {
-        const value = localStorage.getItem(key);
-        if (value) {
-            keysAndValues[key] = value;
-        }
-    });
     const clientId = queryParams.get('client_id') || CLIENT_ID;
+
     const { tool: prevTool } = getLocalStorageValueofClient(`clientID-${clientId}`)
 
     if (prevTool && prevTool !== account) {
-        localStorage.clear();
-
-        Object.keys(keysAndValues).forEach(key => {
-            localStorage.setItem(key, keysAndValues[key]);
-        });
+        handleLocalStorageOnLogout(clientId, true)
     }
 }
 
@@ -177,20 +162,21 @@ const getPlatformImage = () => {
     return platformLogo ?? ''
 }
 
-const handleLocalStorageOnLogout = (currentClientId: string) => {
+const handleLocalStorageOnLogout = (currentClientId: string, isToolChanged = false) => {
     const allKeys = Object.keys(localStorage);
     const keysToKeep = allKeys.filter(key => key.startsWith('clientID') || key.startsWith('streamID'));
 
     const keysAndValues: Record<string, string> = {};
 
-    const { clientId, incentives, links, logo, profileTypeStreamId } = getLocalStorageValueofClient(`clientID-${currentClientId}`)
+    const { clientId, incentives, links, tool, logo, profileTypeStreamId } = getLocalStorageValueofClient(`clientID-${currentClientId}`)
 
     const updatedData = {
         clientId,
         incentives,
         links,
         logo,
-        profileTypeStreamId
+        profileTypeStreamId,
+        ...(isToolChanged ? {} : { tool })
     };
 
     setLocalStorageValue(`clientID-${currentClientId}`, JSON.stringify(updatedData));
@@ -202,11 +188,17 @@ const handleLocalStorageOnLogout = (currentClientId: string) => {
         }
     })
 
-    localStorage.clear();
+    if (!isToolChanged) {
+        localStorage.clear();
+    }
 
     Object.keys(keysAndValues).forEach(key => {
         setLocalStorageValue(key, keysAndValues[key]);
     });
+
+    if (isToolChanged) {
+        window.location.reload();
+    }
 }
 
 const addGlobalLitData = (currentClientId: string) => {
